Functions and type to fit the temperature of blackbody and real surface thermal emission

This Module introduces two main types: - EmPoint for blackbody thermal emission - BandPyrometryPoint for real surface (with spectrally dependent emissivity) This types can be used to solve the least-square optimization problem to fit the temperature and spectral emissivity

BandPyrometry._fill_vander! — Method

_fill_vander!(V, poly_obj::AbstractPolyWrapper,xi)

Function to fill the matrix V columns from polynomial basis functions constructor with argument vector xi

BandPyrometry.bern_max — Method

bern_max(::BernsteinPolyWrapper{D},k::Int)

Returns Bernstein's monomial (maximumvalue, maximumlocation) tuple

BandPyrometry.bern_max_values — Method

bern_max(::VanderMatrix{N,CN,T,NxCN,CNxCN,P}) where {N,CN,T,NxCN,CNxCN,P<:BernsteinPolyWrapper{CN}}

Returns a vector of maximal values of Bernstein basis polynomial basis for particular VanderMatrix

BandPyrometry.denormalize_x — Method

denormalize_x(normalized_x::AbstractVector, x_min,x_max)

Creates normal vector from one created with normalize_x` function

BandPyrometry.disc — Method

disc(x::AbstractVector,bp::BandPyrometryPoint)

Fills emissivity, thermal emission spectrum,evaluates the residuals vector and calculates its norm for the current BandPyrometry point Input: x - optimization variables vector, x=[a1...an,T], where a1...an - emissivity approximations coefficients, T - temperature bp - (modified) current spectral band pytometry point Fills emissivity, thermal emission spectrum,evaluates the residuals vector and calculates its norm for the current BandPyrometry point Input: x - optimization variables vector, x=[a1...an,T], where a1...an - emissivity approximations coefficients, T - temperature bp - (modified) current spectral band pytometry point

BandPyrometry.disc — Method

disc(T,e::EmPoint)

Evaluates the least-square discrepancy between measured and calculates spectra Input: x - optimization variables vector, x=[a1...an,T], where a1...an - emissivity approximations coefficients, T - temperature e - (modified) current bb thermal emission point Evaluates the least-square discrepancy between measured and calculates spectra Input: x - optimization variables vector, x=[a1...an,T], where a1...an - emissivity approximations coefficients, T - temperature e - (modified) current bb thermal emission point

BandPyrometry.em_cons! — Method

em_cons!(constraint_value::AbstractArray,
                        x::AbstractVector, 
                        bp::BandPyrometryPoint)

In-place filling of two-elemnt vector of [minimum,maximum] emissivity in the whole wavelength range This function is used in the constraints Inputs: constraint_value - (modified) two-element vector to be modified in-place x - optimization variables vector, x=[a1...an,T], where a1...an - emissivity approximations coefficients, T - temperature bp - (modified)

BandPyrometry.emissivity! — Method

emissivity!(bp::BandPyrometryPoint,x::AbstractVector)

Fills emissivity for the current BandPyrometry point Input: bp - (modified) current spectral band pytometry point x - optimization variables vector, x=[a1...an,T], where a1...an - emissivity approximations coefficients, T - temperature Fills emissivity for the current BandPyrometry point Input: bp - (modified) current spectral band pytometry point x - optimization variables vector, x=[a1...an,T], where a1...an - emissivity approximations coefficients, T - temperature

BandPyrometry.eval_poly — Method

Function to evaluate polynomials

BandPyrometry.eval_poly — Method

BandPyrometry.evaluate_box_constraints — Method

box_constraints(bp::BandPyrometryPoint)

Evaluates box-constraint of the problem

BandPyrometry.feval! — Method

feval!(bp::BandPyrometryPoint,x::AbstractVector)

Fills both the emissivity and the thermal emission spectrum for the current BandPyrometry point

Input: bp - (modified) current spectral band pytometry point x - optimization variables vector, x=[a1...an,T], where a1...an - emissivity approximations coefficients, T - temperature

BandPyrometry.feval! — Method

feval!(e::EmPoint,t::Float64)

Evaluates bb intensity for temperature t

BandPyrometry.fill_box_constraint! — Method

fill_box_constraint!(lb,ub,::VanderMatrix{N, CN, T, NxCN, CNxCN, P},
            val_bounds::NTuple{2,T}) where {N, CN, T, NxCN, CNxCN, P<:BernsteinSymPolyWrapper}

Evaluates box-boundaries for polynomial coefficients for BernsteinSymPolyWrapper polynomial basis

BandPyrometry.fit_T! — Method

fit_T!(point::Union{EmPoint,BandPyrometryPoint};
        optimizer_name::String="Default",
        is_box_constraint::Bool=false,
        is_lagrange_constraint::Bool=false, 
        emissivity_range::C=nothing, 
        temperature_range::B=nothing) where {B <: Union{AbstractVector,Nothing,NTuple{2}},C <: Union{AbstractVector,Nothing,NTuple{2}}}

Input: point - (modified) real suraface (BandPyrometryPoint) of blackbody thermal emission object (optional) optimizername - the name of optimizer must be the key of optimdic isconstraint - is box-constraint flag islagrange_constraint - is Lagrange-constraint flag Returns: if the point type is EmPoint, the output is: named tuple with (T - fitted temperature, res - optimization output object, optimizer - chosen optimizer) if the point is of the BandPyrometryPoint type, the output is: named tuple with (T - fitted temperature ,a - fitted emissivity approximation coefficients, ϵ - emissivity spectrum in the whole wavelemgth range, res - optimization output object, optimizer - chosen optimizer)

BandPyrometry.fitting_covariance — Method

covariance(bp::BandPyrometryPoint)

Evaluates the covariance matrix as Cov(x) = 2σ²H⁻¹

BandPyrometry.fitting_covariance — Method

fitting_covariance(em::EmPoint{N})

Evaluates the covariance matrix as Cov(x) = 2σ²H⁻¹

BandPyrometry.fitting_variance — Method

fitting_variance(bp::BandPyrometryPoint)

Returns the optimization variable variance (diagonal of the covariance matrix)

BandPyrometry.grad! — Method

grad!(g::AbstractVector,x::AbstractVector,bp::BandPyrometryPoint)

In-place filling of the gradient vector of BandPyrometryPoint at point x Input: x - optimization variables vector, x=[a1...an,T], where a1...an - emissivity approximations coefficients, T - temperature bp - (modified, recalculates residual vector and Jacobian if the currently stored value was obtaibed for another optimization variables array) current spectral band pytometry point In-place filling of the gradient vector of BandPyrometryPoint at point x Input: x - optimization variables vector, x=[a1...an,T], where a1...an - emissivity approximations coefficients, T - temperature bp - (modified, recalculates residual vector and Jacobian if the currently stored value was obtaibed for another optimization variables array) current spectral band pytometry point

BandPyrometry.grad! — Method

grad!(g::AbstractVector,t::Float64 ,e::EmPoint)

In-place filling of the gradient of least-square problem of bb thermal emission spectrum fitting Input: g - gradient vector to be filled t - temperature e - (modified) current bb thermal emission ppoint

BandPyrometry.hess! — Method

hess!(h,x::AbstractVector,bp::BandPyrometryPoint)

In-place filling of the whole hessian matrix for BandPyrometryPoint at point x Input: ha - Hessian matrix to be filled in-place x - optimization variables vector, x=[a1...an,T], where a1...an - emissivity approximations coefficients, T - temperature bp - (modified) current spectral band pytometry point In-place filling of the whole hessian matrix for BandPyrometryPoint at point x Input: ha - Hessian matrix to be filled in-place x - optimization variables vector, x=[a1...an,T], where a1...an - emissivity approximations coefficients, T - temperature bp - (modified) current spectral band pytometry point

BandPyrometry.hess! — Method

hess!(h,t::Float64,e::EmPoint)

In-place filling of least-square problem hessian matrix Input: h - hessian T - temperature e - (modified) current bb thermal emission point

In-place filling of least-square problem hessian matrix Input: h - hessian T - temperature e - (modified) current bb thermal emission point

BandPyrometry.hess_approx! — Method

hess_approx!(ha, x::AbstractVector,bp::BandPyrometryPoint)

In-place filling of the approximate hessian (Hₐ = Jᵀ*J (J - Jacobian)) of BandPyrometryPoint at point vector x, approximate Hessian can be used in optimization methods to approximate the full Hessian (e.g. in Gauss-Newton or Levenberg-Marquardt methods)

Input: ha - Hessian matrix to be filled in-place x - optimization variables vector, x=[a1...an,T], where a1...an - emissivity approximations coefficients, T - temperature bp - (modified) current spectral band pytometry point

BandPyrometry.is_the_same_x — Method

is_the_same_x(v::VanderMatrix,x::AbstractVector)

Checks if input x is the same as the one used for VanderMatrix creation

BandPyrometry.jacobian! — Method

jacobian!(x::AbstractVector,bp::BandPyrometryPoint)

Fills the Jacobian matrix for current bandpyrometry point Input: x - optimization variables vector, x=[a1...an,T], where a1...an - emissivity approximations coefficients, T - temperature bp - (modified, stores Jacobian internally) current spectral band pytometry point Fills the Jacobian matrix for current bandpyrometry point Input: x - optimization variables vector, x=[a1...an,T], where a1...an - emissivity approximations coefficients, T - temperature bp - (modified, stores Jacobian internally) current spectral band pytometry point

BandPyrometry.normalize_x — Method

normalize_x(x::AbstractVector)

Makes all elements of vector x to fit in range -1...1 returns normalized vector , xmin and xmax values All elements of x must be unique Makes all elements of vector x to fit in range -1...1 returns normalized vector , xmin and xmax values All elements of x must be unique

BandPyrometry.optimizer_switch — Method

optimizer_switch(name::String;is_constraint::Bool=false,
            is_lagrange_constraint::Bool=false)

Returns the appropriate optimizer constructor Input: name - the name of th eoptimizer, isconstraint - is box constraint problem formulation, islagrange_constraint - use Lagrange constraints (supported only by IPNewton)

Returns the appropriate optimizer constructor Input: name - the name of th eoptimizer, isconstraint - is box constraint problem formulation, islagrange_constraint - use Lagrange constraints (supported only by IPNewton)

BandPyrometry.polyfit — Method

polyfit(V::VanderMatrix{N,CN,T},x::VT,y::VT) where {N,CN,T<:Number,VT<:Vector{T}}

Fits data x - coordinates, y - values using the VanderMatrix basis function (this coefficients for normalized x-vector)

    (a,y_fitted, gf) = polyfit(V,x,y)
    y_fitted =  V.v*a # V.v - is the vandermonde matrix
    # for normalized x, which has all values within [-1,1] range
    # if a = [a₁ , ..., aₙ], 
    # e.g if V is for standard basis:
    (xnorm,) = normalize_x(x) # returns vector 
    # y_fitted = a₁ + a₂xnorm + ... + aₙxnormⁿ⁻¹

Input: x - coordinates, [Nx0] y - values, [Nx0] returns tuple with vector of polynomial coefficients, values of y_fitted at x points and the norm of goodness of fit

BandPyrometry.polyfitn — Method

polyfitn(V::VanderMatrix{N,CN,T},x::VT,y::VT) where {N,CN,T<:Number,VT<:Vector{T}}

Fits data x - coordinates, y - values using the VanderMatrix basis function (coefficients for unnormalized x-vector)

Input: x - coordinates, [Nx0] y - values, [Nx0] returns tuple with vector of polynomial coefficients, values of y_fitted at x points and the norm of goodness of fit

BandPyrometry.residual! — Method

residual!(bp::BandPyrometryPoint,x::AbstractVector)

Fills emissivity, thermal emission spectrum and evaluates the residuals vector for the current BandPyrometry point

Input: bp - (modified) current spectral band pytometry point x - optimization variables vector, x=[a1...an,T], where a1...an - emissivity approximations coefficients, T - temperature

BandPyrometry.residual! — Method

residual!(e::EmPoint,t::Float64)

Evaluates the residual vector between calculated and measured bb thermal emission intensity spectrum Evaluates the residual vector between calculated and measured bb thermal emission intensity spectrum

BandPyrometry.student_coefficient — Method

student_coefficient(degrees_of_freedom::Int, probability; digits::Int = 3, side::Int = 2)

Evaluates Student's distribution coefficient

BandPyrometry.∇! — Method

∇!(t::Float64,e::EmPoint)

Fills the first derivative of Planck function Input: T - temperature e - (modified) current bb thermal emission point Fills the first derivative of Planck function Input: T - temperature e - (modified) current bb thermal emission point

BandPyrometry.∇²! — Method

∇²!(t::Float64,e::EmPoint)

Fills the second derivative of Planck function Input: T - temperature e - (modified) current bb thermal emission point

Base.:* — Method

*(V::VanderMatrix,a::AbstractVector)

VanderMatrix object can be directly multiplyed by a vector

BandPyrometry.BandPyrometryPoint — Type

BandPyrometryPoint type stores data of thermal emission spectrum of a real body with

emissivity polynomial approximation, and its first and second derivatives it also stores "measurements" vector which further can be fitted, it also provides the constructor BandPyrometryPoint(Imeasured,λ,initialx,polynomialtype) where: - Imeasured is a measured spectrum - λ - wavelength vector (in μm) - initialx - starting parameters vector (initialx[end] - starting temperature, x[1:end-1] - emissivity approximation coefficients) - polynomial_type - string of polynomial (this value governs the Vandermonde matrix form)

N - number ob wavelength points Nx3 - 3N used to store the intermediate data for planck function and derivarives evaluation L - length of optimization variables vector P - number of variables to be optimized (can be less or equal to the number of ) NxP - NP number of elements in jacobian PxP - PP number of elements in hessian Pm1 - P-1 number of parameters approximating NxPm1 - N(P-1) number of vandermatrix elements T - type of data

BandPyrometry.EmPoint — Type

EmPoint type stores data on thermal emission spectrum and its first and second derivatives it also stores "Measurements " vector which further can be fitted? it also provides the constructor EmPoint(Imeasured,λ) - Imeasured is a measured spectrum - λ - wavelength vector (in μm)

BandPyrometry.VanderMatrix — Type

VanderMatrix{M <: SMatrix , R <: SMatrix, V <: SVector}

This type stores the Vandemonde matrix (fundamental matrix of basis functions), supports various types of internal polynomials Structure VanderMatrix has the following fields: v - the matrix itself (each column of this matrix is the value of basis function) vunnorm - version of matrix with unnormalized basis vectors (used for annormalized coefficients of polynomial fitting) xfirst - first element of the initial vector xlast - the last value of the initial vector xi - normalized vector polytype - polynomial type name (nothing depends on this name)

BandPyrometry.VanderMatrix — Method

VanderMatrix(x::StaticArray{Tuple{N},T,1},
                  PolyWrapper::Type{P} # = StandPolyWrapper{CN}
                ) where {N, T, P <:AbstractPolyWrapper{CN,PN}} where {CN,PN}

Input: x - vector of independent variables (coordinates) Val(CN) - vandermatrix column size (degree of polynomial + 1) polytype - polynomial type name, must be member of SUPPORTEDPOLYNOMIAL_TYPES

Core.Union — Method

(poly::Union{LegPolyWrapper,TrigPolyWrapper})(x::Number)

Function returns the summation of polynomials values with coefficients 
LegPolyWrapper wraps the LegendrePolynomials.jl library function to 
make it consistent with Polynomials.jl 
TrigPolyWrapper simple type for trigonometric function polynomils