module with various pyrometers types

Base.isless — Method

Base.isless(p1::Pyrometer,p2::Pyrometer)

Vector of Pyrometer objects can be sorted using isless

Main.BandPyrometry.Pyrometers.fit_ϵ — Method

fit_ϵ(p::Pyrometer,Tmeasured::Float64,Treal::Float64)

Optimizes the emissivity of the pyrometer to make measured by the pyrometer temperature fit
fit the real temperature

Input:
    p - pyrometer object
    Treal - real temperature of the surface, Kelvins
    Tmeasured - temperature measured by the pyrometer, Kelvins

Main.BandPyrometry.Pyrometers.fit_ϵ — Method

fit_ϵ(p::Vector{Pyrometer},Treal::Float64,Tmeasured::Vector{Float64})

Optimizes the emissivity of the pyrometer to make measured by the pyrometer temperature fit
fit the real temperature

Input:
    p - pyrometer objects vector , [Nx0]
    Treal - real temperature of the surface, Kelvins
    Tmeasured - temperatures measured by the pyrometer, Kelvins, [Nx0]

Main.BandPyrometry.Pyrometers.fit_ϵ_wavelength — Method

fit_ϵ_wavelength(p::Pyrometer,Tmeasured::Float64,Treal::Float64)

Fits emissivity and returns it as a vector of the same size as pyrometer's wavelength region
Some pyrometers has 2-wavelength, other work on a single wavelength, for two-wavelength pyrometers
e_out will be a two-element vector
Input:
    p - pyrometer object
    Tmeasured - temperature measured by the pyrometer, Kelvins
    Treal - real temeprature of the surface, Kelvins

Main.BandPyrometry.Pyrometers.full_wavelength_range — Method

full_wavelength_range()

Creates the wavelengths vector covered by the pyrometers_types

Main.BandPyrometry.Pyrometers.produce_pyrometers — Method

produce_pyrometers()

Creates the vector of all supported pyrometers

Main.BandPyrometry.Pyrometers.set_emissivity — Method

set_emissivity(p::Pyrometer,em_value::Float64)

Setter for spectral emissivity

Main.BandPyrometry.Pyrometers.switch_the_type — Method

switch_the_type(λ::Float64)

Returns the type (which can be used as a key of pyrometers_types dict) depending on the wavelengh
Input:
    λ - wavelength in μm

Main.BandPyrometry.Pyrometers.wavelength_number — Method

wavelength_number()

Returns the length of wavelengths vector covered by the pyrometers_types