Functions and type to fit the temperature of blackbody and real surface thermal emission

Base.:* — Method

*(V::VanderMatrix,a::AbstractVector)

VanderMatrix object can be directly multiplyed by a vector

Main.BandPyrometry.Vander — Method

Vander(x::AbstractArray, poly_degree::Number;poly_type::String="stand")

Creates VanderMatrix object of predefied type, all supported polynomial type can 
be found in @supported_polynomial_types

Main.BandPyrometry.box_constraints — Method

box_constraints(bp::BandPyrometryPoint)

Evaluates box-constraint of the problem

Main.BandPyrometry.denormalize_x — Method

denormalize_x(normalized_x::AbstractVector, x_min,x_max)

Creates normal vector from one created with @normalize_x function

Main.BandPyrometry.disc — Method

disc(x::AbstractVector,bp::BandPyrometryPoint)

Fills discrepancy value, bp.e_p strores the residual function norm

Main.BandPyrometry.em_cons! — Method

Evaluates maximum emissivity in the whole wavelength range 
This function is used in the constraints

Main.BandPyrometry.emissivity! — Method

Fills emissivity for the current BandPyrometry point

Main.BandPyrometry.eval_poly — Method

Function to evaluate polynomials

Main.BandPyrometry.feval! — Method

Fills emissivity and emission spectra

Main.BandPyrometry.grad! — Method

grad!(g::AbstractVector,x::AbstractVector,bp::BandPyrometryPoint)

In-place filling of the gradient of BandPyrometryPoint at point x

Main.BandPyrometry.hess! — Method

hess!(h,x::AbstractVector,bp::BandPyrometryPoint)

In-place filling of hessian of BandPyrometryPoint at point x

Main.BandPyrometry.hess_approx! — Method

hess_approx!(ha, x::AbstractVector,bp::BandPyrometryPoint)

In-place filling of the approximate hessian (Hₐ = Jᵀ*J (J - Jacobian)) 
of BandPyrometryPoint at point

Main.BandPyrometry.jacobian! — Method

jacobian!(x::AbstractVector,bp::BandPyrometryPoint)

Fills the Jacobian matrix

Main.BandPyrometry.normalize_x — Method

normalize_x(x::AbstractVector)

Makes all elements of vector x to fit in range -1...1
returns normalized vector , xmin and xmax values
All elements of x must be unique

Main.BandPyrometry.optimizer_switch — Method

optimizer_switch(name::String;is_constraint::Bool=false,
            is_lagrange_constraint::Bool=false)
Returns the appropriate optimizer constructor
Input:
    name - the name of th eoptimizer, 
    is_constraint - is box constraint problem formulation, 
    is_lagrange_constraint - use Lagrange constraints (supported only by IPNewton)

Main.BandPyrometry.polyfit — Method

polyfit(V::VanderMatrix,x::T,y::T) where T<:AbstractVector

Fits data x - coordinates, y - values using the VanderMatrix
basis function

Input:
    x - coordinates, [Nx0]
    y - values, [Nx0]
returns tuple with vector of polynomial coefficints, values of y_fitted at x points
    and the norm of goodness of fit

Main.BandPyrometry.residual! — Method

Fills emissivity, emission spectra and evaluates residual vector

Core.Union — Method

(poly::Union{LegPolyWrapper,TrigPoly})(x::Number)

Function returns the summation of polynomials values with coefficients 
LegPolyWrapper wraps the LegendrePolynomials.jl library function to 
make it consistent with Polynomials.jl 
TrigPoly simple type for trigonometric function polynomils

Main.BandPyrometry.BandPyrometryPoint — Type

    BandPyrometryPoint type stores data of thermal emission spectrum of a real body with 
    emissivity polynomial approximation, and  its first and second derivatives
    it also stores "measurements" vector which further can be fitted, it also 
    provides the constructor BandPyrometryPoint(I_measured,λ,initial_x,polynomial_type) 
    where:
        -  I_measured is a measured spectrum
        -  λ - wavelength vector (in μm)
        - initial_x - starting parameters vector (initial_x[end] - starting temperature,
            x[1:end-1] - emissivity approximation coefficients)
        - polynomial_type - string of polynomial (this value governs the Vandermonde matrix form)

Main.BandPyrometry.EmPoint — Type

EmPoint type stores data on thermal emission spectrum and its first and second derivatives it also stores "Measurements " vector which further can be fitted? it also provides the constructor EmPoint(Imeasured,λ) - Imeasured is a measured spectrum - λ - wavelength vector (in μm)

Main.BandPyrometry.VanderMatrix — Type

VanderMatrix(λ,polynomial_degree,MatrixType,polynomial_type)

This type stores the Vandemonde matrix (fundamental matrix of basis functions),
supports various types of internal polynomials including provided externally
Structure VanderMatrix has the following fields:
    v - the matrix itself (each column of this matrix is the value of basis function)
    x_first -  first element of the initial vector 
    x_last  -  the last value of the initial vector
    xi - normalized vector 
    poly_type  - string of polynomial type name (nothing depends on this name)
    poly_constructor  - the constructor of the polynomial object, accepts the value of 
        polynomial coefficients and returns a callable object to evaluate the obtained 
        polynomial at a praticular point. V-matrix is filled by first creating the polynomial 
        obj with only one non-zero polynomial coefficient and then sending the values of xi
        to the created object